update all samples again

+++ samples/client/petstore/java-helidon-client/v4/mp/src/main/java/org/openapitools/client/model/Order.java
@@ -63,7 +63,7 @@ public enum StatusEnum {
 
 public enum PaymentMethodEnum {
 
-    NUMBER_1(BigDecimal.valueOf(new BigDecimal("1"))), NUMBER_2(BigDecimal.valueOf(new BigDecimal("2")));
+    NUMBER_1(new BigDecimal("1")), NUMBER_2(new BigDecimal("2"));
 
     BigDecimal value;
